+++ tests/models/albef/test_albef.py
@@ -16,7 +16,10 @@ from torchmultimodal.models.albef.model import (
     ALBEFModelWithSimilarity,
     ALBEFSimilarity,
 )
-from torchmultimodal.models.albef.multimodal_encoder import ALBEFMultimodalEncoder
+from torchmultimodal.models.albef.multimodal_encoder import (
+    ALBEFMultimodalEncoder,
+    TransformerCrossAttentionLayer,
+)
 from torchmultimodal.modules.encoders.bert_text_encoder import bert_text_encoder
 from torchmultimodal.utils.common import momentum_update, remove_grad
 
@@ -241,3 +244,58 @@ def test_neg_embeddings(albef_with_sim):
     assert_expected(image_embeds_neg, expected_image_embeds_neg, rtol=0, atol=1e-4)
     assert_expected(text_embeds_neg, expected_text_embeds_neg, rtol=0, atol=1e-4)
     assert_expected(text_atts_neg, expected_text_atts_neg, rtol=0, atol=1e-4)
+
+
+class TestTransformerCrossAttentionLayer:
+    @pytest.fixture(autouse=True)
+    def seed(self):
+        set_rng_seed(4)
+
+    @pytest.fixture
+    def get_encoder_layer(self):
+        def create_layer(norm_first):
+            model = TransformerCrossAttentionLayer(2, 1, 2, norm_first=norm_first)
+            model.eval()
+            return model
+
+        return create_layer
+
+    @pytest.fixture
+    def inputs(self):
+        return torch.randn(1, 2, 2, 2, 2)
+
+    @pytest.fixture
+    def cross_inputs(self):
+        return torch.randn(1, 2, 2, 2, 2)
+
+    def test_forward_prenorm(self, inputs, cross_inputs, get_encoder_layer):
+        model = get_encoder_layer(True)
+        actual = model(inputs, cross_inputs)
+        expected = torch.tensor(
+            [
+                [
+                    [
+                        [[-0.5925, 1.1257], [-0.5925, 1.1257]],
+                        [[-0.5925, 1.1257], [-0.5925, 1.1257]],
+                    ],
+                    [
+                        [[-0.5925, 1.1257], [-0.5925, 1.1257]],
+                        [[-0.5925, 1.1257], [-0.5925, 1.1257]],
+                    ],
+                ]
+            ]
+        )
+        assert_expected(actual, expected, rtol=0, atol=1e-4)
+
+    def test_forward_postnorm(self, inputs, cross_inputs, get_encoder_layer):
+        model = get_encoder_layer(False)
+        actual = model(inputs, cross_inputs)
+        expected = torch.tensor(
+            [
+                [
+                    [[[-1.0, 1.0], [-1.0, 1.0]], [[-1.0, 1.0], [-1.0, 1.0]]],
+                    [[[-1.0, 1.0], [-1.0, 1.0]], [[-1.0, 1.0], [-1.0, 1.0]]],
+                ]
+            ]
+        )
+        assert_expected(actual, expected, rtol=0, atol=1e-4)
